do not hardcode syslog devices
in my specific use case (Debian Jessie), multiple of the candidates
exist, which makes the whole thing *crash* because there are more than
one candidates.

regardless, we shouldn't have to guess where syslog can be: it's the
stdlib's job to do the right thing and find the right device here.

otherwise this should be configurable. so revert part of the previous
commits and restore sanity.