improve irkerd.service
irkerd should not be launched as root.

admins might not remember whether the service is called "irkerd" or
"irker".

"irker daemon" doesn't really describe the service to someone who
doesn't already know what it is.

the .service file does not currently point to the relevant
documentation.

this patch addresses all of the above issues.